drivers: sensor: stm32_temp: attribute for ADC reference voltage added
Added a new attribute SENSOR_ATTR_VREF_MV for setting the ADC reference
voltage after boot time. This was needed for using the driver on a
hardware where the reference voltage is not fixed.

If the new attribute is not set then the ADC reference voltage is used
from the ADC node in the device tree as in older versions of this
driver.

Signed-off-by: Moritz Diller <moritz.diller@toolsforhumanity.com>