drivers: led_strip: apa102: support update channels
apa102 is controlled with 4 bytes, which are received in that order:
brightness (5 bits), blue, green and red.
The rgb api doesn't allow the user to set brightness with the update_rgb
method. Adding update_channels allows full control of the leds in the
led strip.
Based on lpd880x driver

Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>